# refactor(Scalar.AspNetCore): improve GZip check