--- a/.changeset/odd-otters-change.md
+++ b/.changeset/odd-otters-change.md
@@ -1,0 +1,5 @@
+---
+'@scalar/aspnetcore': patch
+---
+
+refactor(Scalar.AspNetCore): Improve Gzip check