--- a/integrations/aspnetcore/tests/Scalar.AspNetCore.Tests/HttpRequestExtensionsTests.cs
+++ b/integrations/aspnetcore/tests/Scalar.AspNetCore.Tests/HttpRequestExtensionsTests.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Scalar.AspNetCore.Tests;
+
+public class HttpRequestExtensionsTests
+{
+    [Theory]
+    [InlineData("gzip, br", true)]
+    [InlineData("br", false)]
+    public void IsGzipAccepted_ShouldReturnTrue_WhenGzipIsPresent(string acceptEncoding, bool expected)
+    {
+        // Arrange
+        var httpRequest = new DefaultHttpContext().Request;
+        httpRequest.Headers.AcceptEncoding = acceptEncoding;
+
+        // Act
+        var result = httpRequest.IsGzipAccepted();
+
+        // Assert
+        result.Should().Be(expected);
+    }
+}