… integrations/aspnetcore/src/Scalar.AspNetCore/Extensions/HttpContextAcceptEncodingCheckExtensions.cs