The performance and complexity of the GZIP check can be simplified. This PR removes some code and improves the performance. I also added dedicated tests for this.

**Benchmark:**

| Method  | Mean      | Error    | StdDev   | Ratio | RatioSD | Allocated | Alloc Ratio |
|-------- |----------:|---------:|---------:|------:|--------:|----------:|------------:|
| Current | 115.61 ns | 2.017 ns | 1.887 ns |  1.00 |    0.02 |         - |          NA |
| New     |  15.13 ns | 0.148 ns | 0.124 ns |  0.13 |    0.00 |         - |          NA |


**Checklist**

I’ve gone through the following:

- [x] I’ve added an explanation _why_ this change is needed.
- [x] I’ve added a changeset (`pnpm changeset`).
- [x] I’ve added tests for the regression or new feature.

<!-- See the [contributing guide](https://github.com/scalar/scalar/blob/main/CONTRIBUTING.md) for more information. -->